+++ accumulo/src/main/scala/geotrellis/spark/io/accumulo/AccumuloCollectionLayerReader.scala
@@ -19,7 +19,9 @@ package geotrellis.spark.io.accumulo
 import geotrellis.spark._
 import geotrellis.spark.io._
 import geotrellis.spark.io.avro._
+import geotrellis.spark.tiling._
 import geotrellis.util._
+import geotrellis.vector._
 
 import org.apache.accumulo.core.data.{Range => AccumuloRange}
 import org.apache.hadoop.io.Text
@@ -30,9 +32,9 @@ import scala.reflect._
 class AccumuloCollectionLayerReader(val attributeStore: AttributeStore)(implicit instance: AccumuloInstance) extends CollectionLayerReader[LayerId] {
 
   def read[
-    K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
+    K: AvroRecordCodec: Boundable: JsonFormat: ClassTag: SpatialComponent,
     V: AvroRecordCodec: ClassTag,
-    M: JsonFormat: GetComponent[?, Bounds[K]]
+    M: JsonFormat: Component[?, Bounds[K]]: Component[?, LayoutDefinition]: Component[?, Extent]
   ](id: LayerId, rasterQuery: LayerQuery[K, M], filterIndexOnly: Boolean) = {
     if (!attributeStore.layerExists(id)) throw new LayerNotFoundError(id)
 
@@ -44,13 +46,16 @@ class AccumuloCollectionLayerReader(val attributeStore: AttributeStore)(implicit
 
     val queryKeyBounds = rasterQuery(metadata)
 
+    val layerMetadata = updateQueriedMetadata[K, M](queryKeyBounds, metadata)
+
     val decompose = (bounds: KeyBounds[K]) =>
       keyIndex.indexRanges(bounds).map { case (min, max) =>
         new AccumuloRange(new Text(AccumuloKeyEncoder.long2Bytes(min)), new Text(AccumuloKeyEncoder.long2Bytes(max)))
       }
 
     val seq = AccumuloCollectionReader.read[K, V](header.tileTable, columnFamily(id), queryKeyBounds, decompose, filterIndexOnly, Some(writerSchema))
-    new ContextCollection(seq, metadata)
+
+    new ContextCollection(seq, layerMetadata)
   }
 }