+++ spark/src/main/scala/geotrellis/spark/io/GenericLayerCopier.scala
@@ -19,8 +19,9 @@ package geotrellis.spark.io
 import geotrellis.spark._
 import geotrellis.spark.io.avro._
 import geotrellis.spark.io.index.KeyIndex
+import geotrellis.spark.tiling._
 import geotrellis.util._
-
+import geotrellis.vector._
 
 import org.apache.avro.Schema
 import org.apache.spark.rdd.RDD
@@ -35,9 +36,9 @@ class GenericLayerCopier[Header: JsonFormat](
 ) extends LayerCopier[LayerId] {
 
   def copy[
-    K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
+    K: AvroRecordCodec: Boundable: JsonFormat: ClassTag: SpatialComponent,
     V: AvroRecordCodec: ClassTag,
-    M: JsonFormat: GetComponent[?, Bounds[K]]
+    M: JsonFormat: Component[?, Bounds[K]]: Component[?, LayoutDefinition]: Component[?, Extent]
   ](from: LayerId, to: LayerId): Unit = {
     if (!attributeStore.layerExists(from)) throw new LayerNotFoundError(from)
     if (attributeStore.layerExists(to)) throw new LayerExistsError(to)